feat: add Java JUnit5 recorder
